+++ e2-mgr-client/src/test/java/org/oransc/ric/portal/dashboard/e2mgr/client/test/E2ManagerClientTest.java
@@ -20,7 +20,7 @@
 package org.oransc.ric.portal.dashboard.e2mgr.client.test;
 
 import org.junit.jupiter.api.Test;
-import org.oransc.ric.e2mgr.client.api.E2ManagerApi;
+import org.oransc.ric.e2mgr.client.api.HealthCheckApi;
 import org.oransc.ric.e2mgr.client.invoker.ApiClient;
 import org.springframework.web.client.RestClientException;
 
@@ -35,9 +35,9 @@ public class E2ManagerClientTest {
 	public void demo() {
 		ApiClient apiClient = new ApiClient();
 		apiClient.setBasePath("http://localhost:30099/");
-		E2ManagerApi e2Api = new E2ManagerApi(apiClient);
+		HealthCheckApi healthApi = new HealthCheckApi(apiClient);
 		try {
-			e2Api.healthCheck();
+			healthApi.healthGet();
 			System.out.println("getHealth answered: " + apiClient.getStatusCode().toString());
 		} catch (RestClientException e) {
 			System.err.println("getHealth failed: " + e.toString());